// src/types/Getters.ts
export interface IPrediction {
    id: number;
    age: number;
    balance: number;
    campaign: number;
    contact_name: string;
    day: number;
    default: boolean;
    duration: number;
    education_level_name: string;
    housing: boolean;
    job_name: string;
    loan: boolean;
    marital_name: string;
    month_name: string;
    pdays: number;
    poutcome_name: string;
    predicted_value: number;
    previous: number;
}